Removed anchor-name cleanup on element
It's unique anyways so cleaning it up shouldn't be a big deal? Also it was breaking anchor positioning if I clicked the same element while the popover was already opened due to a race condition of dismissing the existing bobpop vs calling a new bobpop. In the end I want to be able to let a bobpop be nested within another bobpop if possible (probably using manual as a potential dialog for the parent) so I didn't want to limit that.